Remove unnecesary providers as now we are integrating the app with the backend
