--- src/common/hooks/useInvocation.ts
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import { InvocationContext } from '@/providers/InvocationProvider';
-
-export const useInvocation = () => React.useContext(InvocationContext);
