+++ src/common/types/collection.ts
@@ -1,0 +1,7 @@
+import { Folder } from './folder';
+
+export type Collection = {
+	id: string;
+	name: string;
+	folders: Folder[];
+};
