+++ src/common/types/folder.ts
@@ -1,0 +1,7 @@
+import { Invocation } from './invocation';
+
+export type Folder = {
+	id: string;
+	name: string;
+	invocations: Invocation[];
+};
