+++ src/pages/home/Home.tsx
@@ -1,11 +1,11 @@
+import { useCollectionsQuery } from '@/common/api/collections';
 import CollectionCard from '@/common/components/Collections/CollectionCard';
 import CollectionPlaceholder from '@/common/components/Collections/CollectionPlaceholder';
 import CollectionsEmptyState from '@/common/components/Collections/CollectionsEmptyState';
 import FullscreenLoading from '@/common/views/FullscreenLoading';
-import { useCollections } from '@/providers/CollectionsProvider';
 
 export default function Home() {
-	const { collections, isLoading } = useCollections();
+	const { data, isLoading } = useCollectionsQuery();
 
 	if (isLoading) {
 		return <FullscreenLoading />;
@@ -16,13 +16,13 @@ export default function Home() {
 			<h3 className="text-xl font-bold" data-test="collections-header-title">
 				Collections
 			</h3>
-			{collections.length === 0 ? (
+			{!data || data?.length === 0 ? (
 				<div className="flex justify-center items-center h-full">
 					<CollectionsEmptyState />
 				</div>
 			) : (
 				<div className="flex gap-8 flex-wrap">
-					{collections.map((collection) => (
+					{data.map((collection) => (
 						<CollectionCard
 							key={collection.id}
 							id={collection.id}
